refactor(providers): move providers folder out of components
