+++ apps/ui/src/components/pages/application/form-components/BlockRadioBox.tsx
@@ -21,8 +21,8 @@ import { ColProps, Flex, Form, Radio, Row, theme } from 'antd';
 import { FormItemLayout } from 'antd/es/form/Form';
 import { Controller, ControllerRenderProps, FieldValues, Path, UseControllerProps } from 'react-hook-form';
 
-import { pcglColors } from '@/components/providers/ThemeProvider';
 import { BasicFormFieldProps } from '@/global/types';
+import { pcglColors } from '@/providers/ThemeProvider';
 import { EthicsFileEnum } from '@pcgl-daco/validation';
 
 const { Item } = Form;
