+++ apps/ui/src/index.tsx
@@ -24,10 +24,10 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router';
 
-import ThemeProvider from '@/components/providers/ThemeProvider';
 import AppRouter from '@/pages/AppRouter';
+import ThemeProvider from '@/providers/ThemeProvider';
+import { UserProvider } from '@/providers/UserProvider';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { UserProvider } from './components/providers/UserProvider';
 
 const queryClient = new QueryClient();
