+++ apps/ui/src/pages/AppRouter.tsx
@@ -20,7 +20,6 @@
 import { Navigate, Route, Routes } from 'react-router';
 
 import PageLayout from '@/components/layouts/PageLayout';
-import { ApplicationContextProvider } from '@/components/providers/context/application/ApplicationContextProvider';
 import ApplicationViewer from '@/pages/applications';
 import AccessAgreement from '@/pages/applications/sections/access';
 import Appendices from '@/pages/applications/sections/appendices';
@@ -34,6 +33,7 @@ import RequestedStudy from '@/pages/applications/sections/requestedStudy';
 import DashboardPage from '@/pages/dashboard';
 import HomePage from '@/pages/index';
 import ManageApplicationsPage from '@/pages/manage/applications';
+import { ApplicationContextProvider } from '@/providers/context/application/ApplicationContextProvider';
 import ProtectedRoute from '../components/ProtectedRoute';
 
 export const ApplicationSectionRoutes = [
