+++ apps/ui/src/providers/context/application/ApplicationContextProvider.tsx
@@ -17,43 +17,21 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-/* eslint-disable react-refresh/only-export-components */
+import { createContext, useReducer } from 'react';
 
-import type { UserResponse } from '@pcgl-daco/validation';
-import { createContext, useContext, type PropsWithChildren } from 'react';
-import useGetUser from '../../api/useGetUser';
+import ApplicationReducer from '@/providers/context/application/ApplicationReducer';
+import { type ApplicationContextType, type ApplicationFormState } from '@/providers/context/application/types';
 
-type UserState = {
-	isLoading: boolean;
-	isLoggedIn: boolean;
-	refresh: () => void;
-} & Partial<UserResponse>;
+const initialState: ApplicationFormState = {
+	formState: {
+		isDirty: false,
+	},
+};
 
-const UserContext = createContext<UserState>({ isLoading: true, isLoggedIn: false, refresh: () => {} });
+export const ApplicationContext = createContext<ApplicationContextType | undefined>(undefined);
 
-export function UserProvider({ children }: PropsWithChildren) {
-	const { data, isLoading, refetch } = useGetUser();
+export const ApplicationContextProvider = ({ children }: { children: React.ReactNode }) => {
+	const [state, dispatch] = useReducer(ApplicationReducer, initialState);
 
-	// TODO: update local storage
-
-	const refresh = () => {
-		// TODO: update local storage
-		refetch();
-	};
-
-	const value: UserState = {
-		...data,
-		isLoading,
-		isLoggedIn: isLoading ? false : data ? data.role !== 'ANONYMOUS' : false,
-		refresh,
-	};
-	return <UserContext.Provider value={value}>{children}</UserContext.Provider>;
-}
-
-export function useUserContext() {
-	const context = useContext(UserContext);
-	if (context === undefined) {
-		throw new Error('useCount must be used within a CountProvider');
-	}
-	return context;
-}
+	return <ApplicationContext.Provider value={{ state, dispatch }}>{children}</ApplicationContext.Provider>;
+};
